chore: order 패키지 하위로 4계층 구조 이동
