--- src/main/java/com/yeoljeong/tripmate/application/client/ProductClient.java
@@ -1,9 +1,0 @@
-package com.yeoljeong.tripmate.application.client;
-
-import com.yeoljeong.tripmate.application.dto.command.OrderableProductCommand;
-
-import java.util.UUID;
-
-public interface ProductClient {
-    OrderableProductCommand getProduct(UUID productId, UUID scheduleId);
-}
