--- src/main/java/com/yeoljeong/tripmate/application/dto/command/CreateOrderCommand.java
@@ -1,17 +1,0 @@
-package com.yeoljeong.tripmate.application.dto.command;
-
-import java.util.List;
-import java.util.UUID;
-
-public record CreateOrderCommand(
-        UUID userId,
-        List<OrderItemCommand> orderItems
-) {
-    public record OrderItemCommand(
-            UUID planUnitId,
-            UUID productId,
-            UUID scheduleId,
-            Integer quantity
-    ) {
-    }
-}
