--- src/main/java/com/yeoljeong/tripmate/domain/enums/Country.java
@@ -1,6 +1,0 @@
-package com.yeoljeong.tripmate.domain.enums;
-
-public enum Country {
-    KR,
-    JP
-}
