--- src/main/java/com/yeoljeong/tripmate/domain/enums/OrderStatus.java
@@ -1,7 +1,0 @@
-package com.yeoljeong.tripmate.domain.enums;
-
-public enum OrderStatus {
-    CONFIRMED,
-    COMPLETED,
-    CANCELLED
-}
