--- src/main/java/com/yeoljeong/tripmate/infrastructure/persistence/jpa/OrderJpaRepository.java
@@ -1,10 +1,0 @@
-package com.yeoljeong.tripmate.infrastructure.persistence.jpa;
-
-import com.yeoljeong.tripmate.domain.entity.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface OrderJpaRepository extends JpaRepository<Order, UUID> {
-    boolean existsByUserIdAndOrderItems_PlanUnitId(UUID userId, UUID planUnitId);
-}
