+++ src/main/java/com/yeoljeong/tripmate/order/application/client/ProductClient.java
@@ -1,0 +1,9 @@
+package com.yeoljeong.tripmate.order.application.client;
+
+import com.yeoljeong.tripmate.order.application.dto.command.OrderableProductCommand;
+
+import java.util.UUID;
+
+public interface ProductClient {
+    OrderableProductCommand getProduct(UUID productId, UUID scheduleId);
+}
