+++ src/main/java/com/yeoljeong/tripmate/order/application/dto/command/CreateOrderCommand.java
@@ -1,0 +1,17 @@
+package com.yeoljeong.tripmate.order.application.dto.command;
+
+import java.util.List;
+import java.util.UUID;
+
+public record CreateOrderCommand(
+        UUID userId,
+        List<OrderItemCommand> orderItems
+) {
+    public record OrderItemCommand(
+            UUID planUnitId,
+            UUID productId,
+            UUID scheduleId,
+            Integer quantity
+    ) {
+    }
+}
