+++ src/main/java/com/yeoljeong/tripmate/order/domain/enums/Country.java
@@ -1,0 +1,6 @@
+package com.yeoljeong.tripmate.order.domain.enums;
+
+public enum Country {
+    KR,
+    JP
+}
