+++ src/main/java/com/yeoljeong/tripmate/order/domain/enums/OrderStatus.java
@@ -1,0 +1,7 @@
+package com.yeoljeong.tripmate.order.domain.enums;
+
+public enum OrderStatus {
+    CONFIRMED,
+    COMPLETED,
+    CANCELLED
+}
